requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.3
langchain-chroma>=0.0.3
sentence-transformers
pandas
numpy
scikit-learn
torch
transformers
accelerate
chromadb
regex
openai
tiktoken
requests 
python-dotenv